Rename files Element* to ElementMatrix* to be similar to the name of the class they define. Actually the class name are ElementalMatrix, but there are already something call Elemental, which can cause confusion. There are unchanged files called Element.[ch]pp, the purpose of which I have no idea.
