+++ include/El/core/DistMatrix/CMakeLists.txt
@@ -2,12 +2,12 @@
 set_full_path(THIS_DIR_HEADERS
   AbstractDistMatrix.hpp
   Block.hpp
-  Element.hpp
+  ElementMatrix.hpp
   )
 
 # Add the subdirectories
 add_subdirectory(Block)
-add_subdirectory(Element)
+add_subdirectory(ElementMatrix)
 
 # Propagate the files up the tree
 set(HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}" PARENT_SCOPE)
